dto class packaging
