+++ src/main/java/gugus/pleco/controller/ecodto/RestEcoTimeDto.java
@@ -1,0 +1,12 @@
+package gugus.pleco.controller.ecodto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+import java.time.LocalTime;
+
+@Data
+@AllArgsConstructor
+public class RestEcoTimeDto {
+    LocalTime restCoolTime;
+}
