+++ src/main/java/gugus/pleco/controller/pleedto/CreatePleeDto.java
@@ -1,0 +1,11 @@
+package gugus.pleco.controller.pleedto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class CreatePleeDto {
+    String pleeName;
+    Long completeCount;
+}
