+++ src/main/java/gugus/pleco/controller/pleedto/PerformDto.java
@@ -1,0 +1,11 @@
+package gugus.pleco.controller.pleedto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class PerformDto {
+    String email;
+    String ecoName;
+}
