+++ src/main/java/gugus/pleco/controller/pleedto/PleeDictDto.java
@@ -1,0 +1,10 @@
+package gugus.pleco.controller.pleedto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class PleeDictDto {
+    String pleeName;
+}
