+++ src/main/java/gugus/pleco/controller/pleedto/PleeDto.java
@@ -1,0 +1,11 @@
+package gugus.pleco.controller.pleedto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class PleeDto {
+    String pleeName;
+    Long ecoCount;
+}
