+++ src/main/java/gugus/pleco/controller/userdto/DuplicateDto.java
@@ -1,0 +1,10 @@
+package gugus.pleco.controller.userdto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@AllArgsConstructor
+@Data
+public class DuplicateDto {
+    boolean success;
+}
