+++ src/main/java/gugus/pleco/controller/userdto/LoginDto.java
@@ -1,0 +1,12 @@
+package gugus.pleco.controller.userdto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@AllArgsConstructor
+@Data
+public class LoginDto {
+    String token;
+    boolean success;
+    Long pleeSize;
+}
