+++ src/main/java/gugus/pleco/controller/userdto/SignUpDto.java
@@ -1,0 +1,11 @@
+package gugus.pleco.controller.userdto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@AllArgsConstructor
+@Data
+public class SignUpDto {
+    Long userId;
+    boolean success;
+}
