+++ src/main/java/gugus/pleco/controller/userdto/UserDto.java
@@ -1,0 +1,13 @@
+package gugus.pleco.controller.userdto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class UserDto {
+
+    public String email;
+
+    public String password;
+}
