+++ src/main/java/gugus/pleco/controller/userdto/UserEcoDto.java
@@ -1,0 +1,11 @@
+package gugus.pleco.controller.userdto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class UserEcoDto {
+    String pleeName;
+    String status;
+}
