+++ src/main/java/gugus/pleco/service/UserEcoServiceImpl.java
@@ -1,7 +1,7 @@
 package gugus.pleco.service;
 
 import gugus.pleco.aop.aspect.annotation.Log;
-import gugus.pleco.controller.dto.UserEcoListDto;
+import gugus.pleco.controller.userdto.UserEcoListDto;
 import gugus.pleco.domain.Plee;
 import gugus.pleco.domain.PleeStatus;
 import gugus.pleco.domain.UserEco;
