+++ src/main/java/gugus/pleco/service/UserService.java
@@ -1,10 +1,9 @@
 package gugus.pleco.service;
 
 import gugus.pleco.excetion.UserDupulicatedException;
-import javassist.bytecode.DuplicateMemberException;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.core.userdetails.UserDetailsService;
-import gugus.pleco.controller.dto.UserDto;
+import gugus.pleco.controller.userdto.UserDto;
 import gugus.pleco.domain.User;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
