+++ src/main/java/gugus/pleco/service/UserServiceImpl.java
@@ -13,7 +13,7 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import gugus.pleco.controller.dto.UserDto;
+import gugus.pleco.controller.userdto.UserDto;
 import gugus.pleco.domain.Eco;
 import gugus.pleco.domain.User;
 import gugus.pleco.domain.UserEco;
